+++ src/tests/server.ts
@@ -1137,7 +1137,7 @@ describe('Subscribe', () => {
       });
     });
 
-    // complete
+    // send complete
     client.ws.send(
       stringifyMessage<MessageType.Complete>({
         id: '1',
@@ -1145,20 +1145,14 @@ describe('Subscribe', () => {
       }),
     );
 
-    // confirm complete
-    await client.waitForMessage(({ data }) => {
-      expect(parseMessage(data)).toEqual({
-        id: '1',
-        type: MessageType.Complete,
-      });
-    });
+    await server.waitForComplete();
 
     server.pong();
     server.pong();
     server.pong();
 
     await client.waitForMessage(() => {
-      fail('Shouldnt have received a message');
+      fail("Shouldn't have received a message");
     }, 30);
   });